renames inviteActions to createInviteActions
